refactor: move system indices to lance-table and bloom filter to lance-core
Cleans up the boundary between `lance-index` and `lance-table` in
preparation for moving transaction logic into `lance-table`. Pure
refactor — no behavior change.

Move A: the generic split-block bloom filter primitive (`Sbbf`,
`SbbfBuilder`, `AsBytes`) moves from `lance-index::scalar::bloomfilter`
to `lance-core::utils::bloomfilter`. It is a storage-agnostic data
structure with no Lance semantics, so it belongs in a lower-level crate.
`libm`/`twox-hash` move with it from `lance-index` to `lance-core`.

Move B: the `frag_reuse` and `mem_wal` system indices are table-level
structure persisted as indices — the table format interprets their
contents (fragment remapping, row visibility) rather than treating them
as opaque. Their structs, pb conversions, and logic move to a new
public `lance-table::system_index` module. The `impl Index` adapters
stay in `lance-index` (the `Index` trait is local there), re-exporting
the relocated structs from `lance-table`.

Closes #6987
