+++ rust/lance-core/src/utils.rs
@@ -7,6 +7,7 @@ pub mod assume;
 pub mod backoff;
 pub mod bit;
 pub mod blob;
+pub mod bloomfilter;
 pub mod cpu;
 pub mod deletion;
 pub mod futures;
