+++ rust/lance-index/Cargo.toml
@@ -50,7 +50,6 @@ lance-linalg.workspace = true
 lance-select.workspace = true
 lance-tokenizer.workspace = true
 lance-table.workspace = true
-libm.workspace = true
 log.workspace = true
 ndarray.workspace = true
 num-traits.workspace = true
@@ -70,7 +69,6 @@ crossbeam-queue.workspace = true
 bytes.workspace = true
 chrono.workspace = true
 uuid.workspace = true
-twox-hash = "2.0"
 async-channel = "2.3.1"
 bitpacking = { version = "0.9.2", features = ["bitpacker4x"] }
 rand_distr.workspace = true
