+++ rust/lance-table/src/lib.rs
@@ -5,4 +5,5 @@ pub mod feature_flags;
 pub mod format;
 pub mod io;
 pub mod rowids;
+pub mod system_index;
 pub mod utils;
