+++ rust/lance/src/dataset/mem_wal/memtable.rs
@@ -14,8 +14,8 @@ use std::time::{Duration, Instant};
 use arrow_array::{Array, RecordBatch, RecordBatchIterator};
 use arrow_schema::Schema as ArrowSchema;
 use lance_core::datatypes::Schema;
+use lance_core::utils::bloomfilter::sbbf::Sbbf;
 use lance_core::{Error, Result};
-use lance_index::scalar::bloomfilter::sbbf::Sbbf;
 use tokio::sync::RwLock;
 use tracing::instrument;
 use uuid::Uuid;
