+++ rust/lance/src/dataset/write/merge_insert/inserted_rows.rs
@@ -15,7 +15,7 @@ use arrow_array::{
 use arrow_schema::DataType;
 use deepsize::DeepSizeOf;
 use lance_core::Result;
-use lance_index::scalar::bloomfilter::sbbf::{Sbbf, SbbfBuilder};
+use lance_core::utils::bloomfilter::sbbf::{Sbbf, SbbfBuilder};
 use lance_table::format::pb;
 
 // Default bloom filter config: 8192 items @ 0.00057 fpp -> 16KiB filter
